move: 각 상태 관련 클래스 패키지 domain -> controller로 이동
