+++ src/main/java/chess/controller/command/CommandManager.java
@@ -1,4 +1,4 @@
-package chess.domain.command;
+package chess.controller.command;
 
 import java.util.List;
 import java.util.Map;
@@ -6,11 +6,7 @@ import java.util.function.Consumer;
 
 import chess.controller.dto.GameResultBySideDto;
 import chess.controller.dto.ScoreBySideDto;
-import chess.dao.JdbcChessGameDao;
-import chess.domain.ChessGame;
-import chess.domain.board.Board;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Pieces;
 import chess.domain.position.Position;
 import chess.domain.service.ChessGameService;
