+++ src/main/java/chess/controller/command/PrintGameResult.java
@@ -1,4 +1,4 @@
-package chess.domain.command;
+package chess.controller.command;
 
 import java.util.List;
